Added spectrum 128 tests to run under FUSE

+++ docker/build.sh
@@ -5,19 +5,18 @@ cd "$(dirname "${0}")"
 TAG="$(cat TAG)"
 docker login
 
-# Build a statically linked curl executable, since the regular build has a ton
-# of library dependencies
-rm -rf build
-mkdir build
-cd build
-curl -L 'https://raw.githubusercontent.com/dtschan/curl-static/ab8ccc3ff140af860065c04b1e9bcd20bbe2c2d2/build.sh' > build.sh
-bash build.sh
-cd ..
+if [ ! -d build ]; then
+  # Build a statically linked curl executable, since the regular build has a ton
+  # of library dependencies
+  rm -rf build
+  mkdir build
+  cd build
+  curl -L 'https://raw.githubusercontent.com/dtschan/curl-static/ab8ccc3ff140af860065c04b1e9bcd20bbe2c2d2/build.sh' > build.sh
+  bash build.sh
+  cd ..
+fi
 
 # Build the docker file, containing bash, the cross compile tools, and curl
 docker build --no-cache "-t=${TAG}" .
-
-# Remove the curl build files
-rm -rf build
 
 docker push "${TAG}"
